Update README.md
This project is an alternative for the expensive performance evaluating systems available in the market today. Individual bowlers will be able to evaluate their performances on their own with the help of statistics and graphs provided by our system. More specifically, individuals will be able to keep a check on their consistency as a bowler based on bowling parameters like line, length, and speed.

In this system, we will be developing a grid of velostat polymer which will be used as a sensor for the detection of location of ball’s impact on the pitch. This will help us to provide the line and length of the delivery. 

The data collected from this sensor will be accumulated which can be later used to present to the user using Data Visualization techniques which will make this data intelligible for the user.
The final aim of this project is to provide a performance analyzing and evaluating system which can be utilized by any layman for the betterment of the sport.